split aligner to new file
