+++ config.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Aug 13 19:03:24 2021
+
+@author: fred
+"""
+
+class Config():
+    debug = 0
+    
+config = Config()
